api/requirements.txt
PyJWT
pyotp
APScheduler==3.6.0
asn1crypto==0.24.0
bcrypt==3.1.6
blinker==1.4
cffi==1.12.3
Click==7.0
cryptography==2.7
Flask==1.0.3
Flask-APScheduler==1.11.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SocketIO==4.0.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
passlib==1.7.1
pycparser==2.19
python-dateutil==2.8.0
python-engineio==3.8.1
python-socketio==4.1.0
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
tzlocal==1.5.1
Werkzeug==0.15.4
WTForms==2.2.1
